Create op_amplifier.c
I created a OP amplifier program without going through all the stages required with amplifying a signal in this board. There is some noise but the output is mostly loud and understandable. Noise is present because the signal doesn't go through low pass and high pass filters.
This code is written in C.